logging: switch to slog package (#12)
The standard log package does not provide facilities to implement
different log levels. However, this tool outputs a lot of debug
information, cluttering the logs in a production environment.

Switch to the log/slog package, introduced in Go 1.21, which implements
structured logging and log levels in the Go standard library.

The log level is "info" by default, and configurable by the LOG_LEVEL
environment variable.